requirements.txt
setuptools~=58.1.0
beautifulsoup4~=4.10.0
selenium~=4.1.3
Pillow~=9.0.1
difPy~=2.2
keyboard~=0.13.5